Extract `UnsafeUnbounded`
